Så jasmine kan se

+++ domain/src/Solution.jl
@@ -10,7 +10,7 @@ mutable struct Solution
     nTaxi::Int
     totalRideTime::Int
     totalViolationTW::Int
-    totalDistance::Float32
+    totalDistance::Int
     idleTime::Int
 
 end
